chore: move provider wrappers to separate files
