+++ src/pages/ProjectOverviewPage/ProjectOverviewTableProvider.tsx
@@ -1,0 +1,48 @@
+import { FC } from 'react'
+import {
+  ProjectTableProvider,
+  SelectionCellsProvider,
+  SelectedRowsProvider,
+  CellEditingProvider,
+  useProjectTableModuleContext,
+} from '@shared/containers/ProjectTreeTable'
+import { NewEntityProvider } from '@context/NewEntityContext'
+import { usePowerpack } from '@shared/context'
+import { useProjectOverviewContext } from './context/ProjectOverviewContext'
+import { ProjectTableQueriesProvider } from '@shared/containers/ProjectTreeTable/context/ProjectTableQueriesContext'
+import useTableQueriesHelper from './hooks/useTableQueriesHelper'
+import ProjectOverviewPage from './ProjectOverviewPage'
+
+const ProjectOverviewTableProvider: FC = () => {
+  const props = useProjectOverviewContext()
+  const modules = useProjectTableModuleContext()
+
+  const { updateEntities, getFoldersTasks } = useTableQueriesHelper({
+    projectName: props.projectName,
+  })
+
+  const powerpack = usePowerpack()
+
+  return (
+    <ProjectTableQueriesProvider {...{ updateEntities, getFoldersTasks }}>
+      <ProjectTableProvider
+        {...props}
+        powerpack={powerpack}
+        modules={modules}
+        groupByConfig={{ entityType: 'task' }}
+      >
+        <NewEntityProvider>
+          <SelectionCellsProvider>
+            <SelectedRowsProvider>
+              <CellEditingProvider>
+                <ProjectOverviewPage />
+              </CellEditingProvider>
+            </SelectedRowsProvider>
+          </SelectionCellsProvider>
+        </NewEntityProvider>
+      </ProjectTableProvider>
+    </ProjectTableQueriesProvider>
+  )
+}
+
+export default ProjectOverviewTableProvider
